Created CanvasObject as base class for all other CavasX types
