+++ lib/models/canvas_user.rb
@@ -1,4 +1,4 @@
-class CanvasUser < Forgery
+class CanvasUser < CanvasObject
   attr_reader :name, :sis_id, :login_id, :email, :time_zone
 
   def to_s
@@ -55,30 +55,6 @@ class CanvasUser < Forgery
       end
     end
     return users
-  end
-
-  def self.push_csv_to_canvas(opts={})
-    if(opts.nil? || opts[:host].nil? || opts[:token].nil?)
-      raise 'Please provide options hash with both :host and :token'
-    end
-    if(!File.file?("./users.csv"))
-      raise 'Please gen_file before trying to push to canvas'
-    end
-    uri = URI.parse("https://#{opts[:host]}/api/v1/accounts/1/sis_imports.json?import_type=instructure_csv")
-    request = Net::HTTP::Post.new(uri)
-    request.content_type = "text/csv"
-    request["Authorization"] = "Bearer #{opts[:token]}"
-    request.body = ""
-    request.body << File.read("users.csv")
-
-    req_options = {
-      use_ssl: uri.scheme == "https",
-    }
-
-    response = Net::HTTP.start(uri.hostname, uri.port, req_options) do |http|
-    http.request(request)
-  end
-
   end
 
 end
